+++ src/main/java/refinedstorage/api/autocrafting/registry/ICraftingTaskFactory.java
@@ -17,10 +17,11 @@ public interface ICraftingTaskFactory {
      * Returns a crafting task for a given NBT tag and pattern.
      *
      * @param world   the world
+     * @param depth   the depth of the crafting task to create
      * @param tag     the NBT tag, if this is null it isn't reading from disk but is used for making a task on demand
      * @param pattern the pattern
      * @return the crafting task
      */
     @Nonnull
-    ICraftingTask create(World world, @Nullable NBTTagCompound tag, ICraftingPattern pattern);
+    ICraftingTask create(World world, int depth, @Nullable NBTTagCompound tag, ICraftingPattern pattern);
 }