+++ src/main/java/refinedstorage/api/network/NetworkUtils.java
@@ -27,8 +27,8 @@ public final class NetworkUtils {
         return network.getPattern(stack, CompareUtils.COMPARE_DAMAGE | CompareUtils.COMPARE_NBT);
     }
 
-    public static ICraftingTask createCraftingTask(INetworkMaster network, ICraftingPattern pattern) {
-        return RefinedStorageAPI.instance().getCraftingTaskRegistry().getFactory(pattern.getId()).create(network.getNetworkWorld(), null, pattern);
+    public static ICraftingTask createCraftingTask(INetworkMaster network, int depth, ICraftingPattern pattern) {
+        return RefinedStorageAPI.instance().getCraftingTaskRegistry().getFactory(pattern.getId()).create(network.getNetworkWorld(), depth, null, pattern);
     }
 
     public static boolean hasPattern(INetworkMaster network, ItemStack stack) {
@@ -62,7 +62,7 @@ public final class NetworkUtils {
             ICraftingPattern pattern = network.getPattern(stack, compare);
 
             if (pattern != null) {
-                network.addCraftingTask(createCraftingTask(network, pattern));
+                network.addCraftingTask(createCraftingTask(network, 0, pattern));
             }
         }
     }