+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.0.6
 - Added crafting preview screen (way2muchnoise)
+- Added max crafting task depth (raoulvdberge)
 
 ### 1.0.5
 - Fixed crafting a complex item causes the process to flow off the Crafting Monitor's GUI (raoulvdberge)